+++ service/cognitoidentityprovider/api_op_AdminRespondToAuthChallenge.go
@@ -281,36 +281,17 @@ type AdminRespondToAuthChallengeInput struct {
 	ChallengeResponses map[string]string
 
 	// A map of custom key-value pairs that you can provide as input for any custom
-	// workflows that this action triggers.
-	//
-	// You create custom workflows by assigning Lambda functions to user pool
-	// triggers. When you use the AdminRespondToAuthChallenge API action, Amazon
-	// Cognito invokes any functions that you have assigned to the following triggers:
-	//
-	//   - Pre sign-up
-	//
-	//   - custom message
-	//
-	//   - Post authentication
-	//
-	//   - User migration
-	//
-	//   - Pre token generation
-	//
-	//   - Define auth challenge
-	//
-	//   - Create auth challenge
-	//
-	//   - Verify auth challenge response
+	// workflows that this action triggers. You create custom workflows by assigning
+	// Lambda functions to user pool triggers.
 	//
 	// When Amazon Cognito invokes any of these functions, it passes a JSON payload,
 	// which the function receives as input. This payload contains a clientMetadata
 	// attribute that provides the data that you assigned to the ClientMetadata
-	// parameter in your AdminRespondToAuthChallenge request. In your function code in
-	// Lambda, you can process the clientMetadata value to enhance your workflow for
-	// your specific needs.
+	// parameter in your request. In your function code, you can process the
+	// clientMetadata value to enhance your workflow for your specific needs.
 	//
-	// For more information, see [Using Lambda triggers] in the Amazon Cognito Developer Guide.
+	// To review the Lambda trigger types that Amazon Cognito invokes at runtime with
+	// API requests, see [Connecting API actions to Lambda triggers]in the Amazon Cognito Developer Guide.
 	//
 	// When you use the ClientMetadata parameter, note that Amazon Cognito won't do
 	// the following:
@@ -325,7 +306,7 @@ type AdminRespondToAuthChallengeInput struct {
 	//   - Encrypt the ClientMetadata value. Don't send sensitive information in this
 	//   parameter.
 	//
-	// [Using Lambda triggers]: https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-identity-pools-working-with-aws-lambda-triggers.html
+	// [Connecting API actions to Lambda triggers]: https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-pools-working-with-lambda-triggers.html#lambda-triggers-by-event
 	ClientMetadata map[string]string
 
 	// Contextual data about your user session like the device fingerprint, IP