+++ service/cognitoidentityprovider/api_op_ConfirmForgotPassword.go
@@ -74,19 +74,17 @@ type ConfirmForgotPasswordInput struct {
 	AnalyticsMetadata *types.AnalyticsMetadataType
 
 	// A map of custom key-value pairs that you can provide as input for any custom
-	// workflows that this action triggers.
+	// workflows that this action triggers. You create custom workflows by assigning
+	// Lambda functions to user pool triggers.
 	//
-	// You create custom workflows by assigning Lambda functions to user pool
-	// triggers. When you use the ConfirmForgotPassword API action, Amazon Cognito
-	// invokes the function that is assigned to the post confirmation trigger. When
-	// Amazon Cognito invokes this function, it passes a JSON payload, which the
-	// function receives as input. This payload contains a clientMetadata attribute,
-	// which provides the data that you assigned to the ClientMetadata parameter in
-	// your ConfirmForgotPassword request. In your function code in Lambda, you can
-	// process the clientMetadata value to enhance your workflow for your specific
-	// needs.
+	// When Amazon Cognito invokes any of these functions, it passes a JSON payload,
+	// which the function receives as input. This payload contains a clientMetadata
+	// attribute that provides the data that you assigned to the ClientMetadata
+	// parameter in your request. In your function code, you can process the
+	// clientMetadata value to enhance your workflow for your specific needs.
 	//
-	// For more information, see [Using Lambda triggers] in the Amazon Cognito Developer Guide.
+	// To review the Lambda trigger types that Amazon Cognito invokes at runtime with
+	// API requests, see [Connecting API actions to Lambda triggers]in the Amazon Cognito Developer Guide.
 	//
 	// When you use the ClientMetadata parameter, note that Amazon Cognito won't do
 	// the following:
@@ -101,7 +99,7 @@ type ConfirmForgotPasswordInput struct {
 	//   - Encrypt the ClientMetadata value. Don't send sensitive information in this
 	//   parameter.
 	//
-	// [Using Lambda triggers]: https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-identity-pools-working-with-aws-lambda-triggers.html
+	// [Connecting API actions to Lambda triggers]: https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-pools-working-with-lambda-triggers.html#lambda-triggers-by-event
 	ClientMetadata map[string]string
 
 	// A keyed-hash message authentication code (HMAC) calculated using the secret key