+++ service/keyspacesstreams/types/errors.go
@@ -106,6 +106,10 @@ func (e *ResourceNotFoundException) ErrorFault() smithy.ErrorFault { return smit
 // time. Implement exponential backoff in your retry strategy to handle this
 // exception. Reducing your request frequency or distributing requests more evenly
 // can help avoid throughput exceptions.
+//
+// This exception can also occur when more than two processes are reading from the
+// same stream shard at the same time. Ensure that only one process reads from a
+// stream shard at the same time.
 type ThrottlingException struct {
 	Message *string